refactor: Separate the writer logic
