+++ src/cmd.rs
@@ -1,4 +1,5 @@
 pub mod aws_detect;
+pub mod aws_detect_writer;
 pub mod aws_metrics;
 pub mod aws_summary;
 pub mod update;
