+++ src/cmd/aws_summary.rs
@@ -252,7 +252,7 @@ pub fn aws_summary(
     };
     let abused_aws_api_values: Vec<String> = abused_aws_api_calls.values().cloned().collect();
     if let Some(d) = directory {
-        process_events_from_dir(summary_func, d, true, no_color).unwrap();
+        process_events_from_dir(summary_func, d, true, &no_color).unwrap();
         output_summary(
             &user_data,
             output,
